requirements.txt
pathlib==1.0.1
pytest-shutil==1.7.0
openpyxl==3.1.2
python-docx==0.8.11
PyPDF2==3.0.1
spacy==3.6.0
matplotlib==3.7.2
pyautogui==0.9.54
requests==2.31.0
pandas==2.0.3
pillow==10.0.0
numpy==1.23.1
tensorflow==2.13.0
imageai~=3.0.3
TorchVision==0.15.2
torch>=1.9.0
opencv-python>=4.1.2
--extra-index-url https://download.pytorch.org/whl/cpu
--extra-index-url https://download.pytorch.org/whl/cpu
pytest==7.1.3
tqdm==4.64.1
scipy>=1.7.3
mock==4.0.3